[Fix] #1018 SonarCloud 잔여 지적 반영 — cross-file 중복·MINOR push 정리 (Refs #1018)
SonarCloud Quality Gate가 new_duplicated_lines_density 3.1%(>3%)로 여전히 fail이었다.
API로 실측(component_tree, duplications/show)한 결과 중복 상대는 sibling producer였다.

- 실측: build-privacy-consistency-evidence.mjs L349-365(17줄)와
  build-mobile-consumption-evidence.mjs L263-279(17줄)가 candidate-context CLI
  진입점 보일러플레이트(argument 파서 + entry 판별 + candidate-context/output 필수
  인자 검증 + repo-root/provenance 파싱 + build 호출 + mkdirSync/writeFileSync)로
  완전히 동일했다. build-planner-success-evidence.mjs는 --canary-result 등 추가
  인자가 있어 이 shape과 다르므로 대상에서 제외.
- tools/release/lib/candidate-context-evidence-cli.mjs 신규(tools/datapack/lib/
  관례를 따름): 두 producer가 공유하는 최소 CLI 진입점을
  runCandidateContextEvidenceCli(importMetaUrl, build)로 추출했다. 두 producer
  모두 동작 변경 없이 이 helper 호출로 대체(추출-대체만, 로직 재구현 없음).
  CLI 실행·필수 인자 누락 에러 경로를 재실행해 이전과 동일한 결과·에러 메시지를
  확인했다.
- MINOR 2건(javascript:S7778, "Do not call Array#push() multiple times",
  L233/234): evaluateAnswerMatrixConsistency의 연쇄 push() 3~4회를 matrix.flatMap +
  스프레드로 재작성해 제거했다. entry 순서·판정 순서는 이전과 동일하게 유지했다
  (테스트가 검증하는 순서 무관 assertion과도 무관하게 실제 순서도 보존).

검증: build-privacy-consistency-evidence.test.mjs(7)+build-mobile-consumption-evidence.test.mjs+
build-planner-success-evidence.test.mjs 합계 28/28 pass, repository-contract.test.mjs
218/219 pass(잔여 1건은 base 기존 실패로 무관), 두 producer CLI 재실행으로 SATISFIED
판정과 필수 인자 누락 에러 경로 불변 확인.
